blkback: collect per-ring occupancy statistics
It's useful to see the ring queue depth (would a larger ring help?)

Note our method of sampling the depth (counting the number of
packets retrieved) results in 'virtual' depths larger than the
physical size of the ring, as packets can (ideally should) be added
while we're removing them. If we see large 'virtual' depths then
we should (a) celebrate; and (b) not bother enlarging the physical
ring size.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>