Serialise calls to Block.connect
Before this patch it was possible for two threads to both want a
connection to the same disk, and for the two calls to `connect`
to call `plug` concurrently on the same device. This patch serialises
all `Block.connect` calls, which is a bit heavyweight but it should
be safe.

Fixes #31

Signed-off-by: David Scott <dave@recoil.org>